idiot

+++ includes/shader.h
@@ -22,7 +22,7 @@ public:
 
     void setMat4(const std::string& name, const glm::mat4& mat) const;
     void setInt(const std::string& name, int value) const;
-    void setFloat(const std::string& name, int value) const;
+    void setFloat(const std::string& name, float value) const;
 
 private: 
     void check_compile(GLuint shader);
